+++ library/src/main/kotlin/org/coner/crispyfish/filetype/registration/RegistrationColumns.kt
@@ -7,6 +7,7 @@ sealed class RegistrationColumns(val heading: String, var index: Int? = null) {
     class LastName : RegistrationColumns("Last Name")
     class CarModel : RegistrationColumns("Car Model")
     class CarColor : RegistrationColumns("Car Color")
+    class MemberNumber : RegistrationColumns("Member #")
     class RawResultTime : RegistrationColumns("Raw Time")
     class RawResultPosition : RegistrationColumns("Raw Pos.")
     class PaxResultTime : RegistrationColumns("Pax Time")